socket improvements
